client/requirements.txt
colorama==0.4.6
colorlog==6.8.2
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pandas==2.2.2
patsy==0.5.6
pillow==10.3.0
PyAudio==0.2.14
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
scipy==1.13.0
six==1.16.0
statsmodels==0.14.1
tzdata==2024.1
